requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiofiles
python-multipart
requests
Pillow
